attempting change

+++ ncc-packer/build.gradle.kts
@@ -4,18 +4,19 @@ plugins {
     kotlin("js")
 }
 
-val tmpDir = "${rootProject.buildDir}/customTmp"
+val tmpDir = "${rootProject.rootDir}/dist-temp"
 
 val copyDir = tasks.register("copyDir") {
     val inputDir = "${rootProject.buildDir}/compileSync/main/productionExecutable/kotlin/"
-    val outputDir = "$tmpDir/indexDir"
-    inputs.dir(inputDir)
-    outputs.dir(outputDir)
-    outputs.upToDateWhen { true }
-    outputs.cacheIf { true }
+    val outputDir = tmpDir
+//    inputs.dir(inputDir)
+//    outputs.dir(outputDir)
+//    outputs.upToDateWhen { true }
+//    outputs.cacheIf { true }
     doLast {
         val output = File(outputDir)
-        output.deleteRecursively()
+        if (output.exists())
+            output.deleteRecursively()
         File(inputDir).copyRecursively(output, overwrite = true)
         val originIndex = File("$outputDir/Max-Commit-Count.js")
         originIndex.copyTo(File("$outputDir/index.js"))
@@ -25,11 +26,11 @@ val copyDir = tasks.register("copyDir") {
 
 val copyNodes = tasks.register("copyNodes") {
     val inputDir = "${rootProject.buildDir}/js/node_modules"
-    val outputDir = "$tmpDir/node_modules"
-    inputs.dir(inputDir)
-    outputs.dir(outputDir)
-    outputs.upToDateWhen { true }
-    outputs.cacheIf { true }
+    val outputDir = "${rootProject.rootDir}/node_modules"
+//    inputs.dir(inputDir)
+//    outputs.dir(outputDir)
+//    outputs.upToDateWhen { true }
+//    outputs.cacheIf { true }
     doLast {
         val output = File(outputDir)
         output.deleteRecursively()
@@ -52,7 +53,7 @@ kotlin {
                 outputs.upToDateWhen { true }
                 outputs.cacheIf { true }
                 args(
-                    "$inputFile/indexDir/index.js",
+                    "$inputFile/index.js",
                     outputDir
                 )
             }
